Moved the implementation of HQRRP into the header. The build currently fails due to messy include statements. After rebase and minor changes, i expect it to be fine.
